Skillbridge/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
SQLAlchemy==2.0.40
cryptography==36.0.2
Flask-Cors==4.0.0
PyJWT==2.8.0
gunicorn==21.2.0
